chakrashim: fix toString on cross context objects
Object.prototype.toString called on a cross context object will either
return [object Object] or [object Function].  This is because cross
context object using the chakra shim is a proxy.  We need to replace
the default Object.prototype.toString with our shimed version that
will unwrap the proxy and call the target context's
Object.prototype.toString.

Also if the target cross context object is an array, create an array
as the fakeTargetObject so that Array.isArray API return true correctly.

This fixes some failing cross realm toString unittest in lodash.