chakrashim: fix some cross-context bugs
Change ContextShim::RegisterCrossContextObject() to use a finalizer
instead of JsSetObjectBeforeCollectCallback. The later callback does
not guarantee object collection. The object can be subsequently revived
as a result of other objects being revived. When this happens we would
clear the registered cross-context mapping info prematurely.

Change UnwrapIfCrossContext() to use JsGetOwnPropertyDescriptor to read
CrossContextTargetSymbol property. Original JsGetProperty could read
from a prototype object and incorrectly unwraps the prototype object
instead of the current object.

Change MarshalObjectToContext() to re-enter correct fromContext when the
object was cross-context and unwrapped.

Change GetOwnPropertyDescriptorCallback() to return "undefined" instead
of "false" for missing cases. Returning "false" is wrong.

Reviewed-by: kunalspathak, curtisman