gyp: add msvs_windows_target_platform_version support
VS2015 requires WindowsTargetPlatformVersion being specified otherwise
it falls back to WindowsSdk 8.1. Add gyp support to populate VS global
property WindowsTargetPlatformVersion when gyp
msvs_windows_target_platform_version property is specified. If it is a
version number, use as is. Otherwise use it as a key to query the
registry and find the latest installed sdk.

Reviewed-by: munyirik