chakrashim: isMapIterator/isSetIterator bug fix
We added special symbol to detect  mapIterator and setIterator.
IsMapIterator and IsSetIterator extracts the value of the symbol to
check if the object is map or set iterator. However built in objects like
'process' has native getter set that converts the lookup properties to
string. Symbols throws type error for implicit toString conversion and
hence the error.

I changed the symbol key to string key and verified unit test.
Morever string properties will be helpful when we clone objects in cross
context scenarios since we iterate on getOwnPropertyNames and not
on getOwnPropertySymbols (utils.cloneObject method).

Also included chakra_shim.js in jslint.

Reviewed-By: @JianChun