deps: Modified chakracore to enable/disable SIMD
Currently chakracore has `SIMD` under experimental flag which is
ON by default. However recently v8 added this feature under
experimental mode. This fails basic scenario of `util-inspect`.
Temporarily addressed this by exposing a flag in `chakracore`
that will control if `SIMD` will be present on global object or not.
Long time solution would be to expose config flags to JSRT layer
so that ES6 experimental features can be turned ON/OFF on demand.

Opened https://github.com/Microsoft/ChakraCore/issues/1064 to track this
in chakracore.

PR-URL: nodejs/node-chakracore#74
Reviewed-By: Jianchun Xu <Jianchun.Xu@microsoft.com>