test: Suppress debugger test and fixed jslint
* Suppress running debugger test
* Fixed jslint warning that came post-merge
* Currently there are 18 failures. `parallel/test-util-inspect-simd`
fails because chakracore has not implemented all the SIMD objects
that test uses. `parallel/test-vm-low-stack-space` is also a new
failure that expects vm to run script after we catch `out of stack
space` exception. I will debug this further to see gap in chakrashim
but for now, letting this test to fail with node+chakracore.

PR-URL: nodejs/node-chakracore#74
Reviewed-By: Jianchun Xu <Jianchun.Xu@microsoft.com>