src,build: added todo for isolate_data
isolate_data is created before setting the context. However
it fails for chakracore because chakrashim calls `JsPointerToString`
internally to populate properties for isolate_data. For now I have
add `#ifdef` to set context before creating isolate_data. But this
need to be fixed in future.

vcbuild.bat passes `--without-inspector` if engine is chakracore

PR-URL: nodejs/node-chakracore#79
Reviewed-By: Jianchun Xu <Jianchun.Xu@microsoft.com>