chakrashim: fix weak persistent dtors on dispose
node-v8 does not run destructors for weak persistents during
Isolate::Dispose, but node-chakracore did.  This is problematic because
AsyncWrap::~AsyncWrap accesses the node::Environment, which is out of
scope by the time Isolate::Dispose is called.  Recent refactoring to
node::Start() revealed this existing issue by changing the stack frame
that contains Environment to be gone once Isolate::Dispose is called.
Previously we got by by accident since the Environment memory was still
intact during Dispose.

Fix is to not call weak persistent callback if the current IsolateShim
is disposing at the time of callback.

Fixes #165