chakrashim, test: skip a buffer test due to bug
ChakraCore has a TypedArray JIT bug where accessing .length always gives
you the buffer's length, rather than the own property if one exists, in
JIT optimized code.  See
https://github.com/Microsoft/ChakraCore/issues/2319

This is a corner case bug since this it is very uncommon to override the
length property of a TypedArray.  Until the bug is fixed we'll leave the
affected node test disabled when run on node-chakracore.