ios: Correct overflow in ChakraCore's Heap Blocks
The objectCount calculation in the HeapBlock constructors
casts the result of multiplying the PageCount with the PageSize
to an unsigned short. PageCount is defined as 4 for 64 bit
architectures and on iOS the PageSize is 16KB. 4*16KB=64KB , which
overflows an unsigned short and wraps around its value to 0.
Correct this by casting to unsigned short only after dividing
the result of the mentioned multiplication. The assert that precedes
the changed lines seems to indicate this was the intended behaviour.