[Merge Microsoft/Chakracore@45b82ed194] [1.6>master] [MERGE #2991 @pleath] Fix perf regression in typed array tests.
Merge pull request #2991 from pleath:2945

If we bail out on array access because we needed a helper call to, for instance, convert the value stored to a typed array, then when we interpret the array access the profile info needs to be updated to indicate that the helper call was needed. Otherwise, on re-jit, we'll generate the same code again, and we'll keep bailing out. Try to be proactive in the case of non-number value stored to a typed array, but also make use of the fact that we had BailOutOnArrayAccessHelperCall, since helper call cases can't be perfectly predicted in the interpreter.