n-api: enable napi_wrap() to work with any object
Change to inserting an external object into the wrapper object's
prototype chain, instead of injecting an alternate `this` object
in the constructor callback adapter. The latter approach didn't
work in certain scenarios because the JSRT runtime still returned
the original `this` object.

And with this change, the setting and checking of the extension
flag, which distinguished wrapper objects from external objects,
can be removed.

Also removing the CallbackInfo.returnValue field which is not
used anymore (since we switched to direct return values).

PR-URL: https://github.com/nodejs/node-chakracore/pull/269
Reviewed-By: Taylor Woll <taylor.woll@microsoft.com>
Reviewed-By: Kyle Farnung <kfarnung@microsoft.com>