[Merge Microsoft/Chakracore@5026351659] [1.6>master] [MERGE #3075 @Cellule] WASM: Prevent folding cmp instr
Merge pull request #3075 from Cellule:wasm/x86_cmp_swap

If we run out of registers during regalloc we'll try to fold the instruction and sometimes swap the operands.
In the case of `cmp` we swap the operands if we're able to reverse the logic of the next branch.

However, it is possible that another instr will want to use the flags set by `cmp` other than the branch.