[Merge Microsoft/Chakracore@4280fb19ba] [1.6>master] [MERGE #3048 @MikeHolman] fix issue with rejit exception and no profile info
Merge pull request #3048 from MikeHolman:rejitexception

In some cases it's possible to have a rejit exception which wants to disable some optimization even if you don't have profile info. For this to work it required refactoring a bit in how we check if an optimization is disabled, and while I was here I decided to do a bit of cleanup around the RejitReason enum as well.