[Merge Microsoft/Chakracore@1d911f4905] [MERGE #3160 @digitalinfinity] Add event tracing function in test build
Merge pull request #3160 from digitalinfinity:event_trace

This commit introduces a global function chWriteTraceEvent. This function is analogous to msWriteProfilerMark in IE/Edge- however, it's only exposed in debug and test builds of Chakra. It's intention is for use in perf investigations.

The code to that actually fires the ETW event is in PlatformAgnostic. Currently it behaves as a no-op on Linux but it's intended to be filled out once we choose a tracing mechanism for Linux.