[Merge Microsoft/Chakracore@51df492ce6] [1.6>master] [MERGE #3172 @leirocks] fix etw logging for very long name function loop body
Merge pull request #3172 from leirocks:auxptr_etw_loop

if a function has very long name(longer than 256), we calculate the loop body name from a loop number, the loop number is calculated from loop header, which is not necessary because after loop body code gen we already know the loop number. while calculating the loop number we access auxPtrs to get the loop header array in background thread, which is not allowed after JIT on the function is started.

the fix is to just not calculate the loop number again.