[Merge Microsoft/Chakracore@0da0541286] [1.7>master] [1.6>1.7] [MERGE #3117 @meg-gupta] Fix region propagation of finally end blocks
Merge pull request #3117 from meg-gupta:tryfinallyfixes

We add edge from infinite loops in finally region to end of finally block, so that unreachable block elimination does not remove finally end block. We have to retain finally end block, so that we can add edge from finally end block to early exit.

When the only edge to finally block is an edge from a loop within some child region, we can end up in propagating incorrect region info from the predecessor. We have to use the region on the finally label for such an end of finally block.

So create a new map called leaveNullLabelToFinallyLabelMap and pick up region info from here.

Also piggybacking other small changes:
- dumping
- fix asserts for Br instructions :
Leaves in non exception finally are converted to Br instructions, for such branches the target region is the parent of the current region.
Also, a BrOnException from finally to early exit can be converted to BrOnNoException and Br, we cannot assert for such a Br that the target region is same as Br region