[Merge Microsoft/Chakracore@66bc22eceb] [1.6>1.7] [MERGE #3382 @rajatd] On .caller, only invalidate cached scopes for nested parents
Merge pull request #3382 from rajatd:cachedScopeStackWalk

When a function escapes as the return value of .caller, we currently walk the whole stack and invalidate cached scope on every function on the stack that has one set. However, it seems that we only need to do the invalidation for nested parents of the result of .caller. This PR attempts to do just that.