chakrashim: fix console autocomplete in VS Code
The autocomplete for VS Code (and probably Chrome) sends a blob of
JavaScript to the engine which enumerates/filters the properties and
returns them. The code wasn't respecting the `returnByValue` parameter
and VS Code was unable to get the length of the undefined `value` field
in the response.

While I was in there I also removed some unneeded fields from the
`JavaScriptCallFrame` class and fixed a build warning (C4800) in v8.h.

PR-URL: https://github.com/nodejs/node-chakracore/pull/356
Reviewed-By: Jimmy Thomson <jithomso@microsoft.com>