chakrashim: removing ObjectPrototypeToStringShim
To support usage of `v8::FunctionTemplate::className` chakrashim provided
a custom Object.prototype.toString that would require JSRT call out
into an external function, and then most of the time return back into
JSRT to use the defualt function.

The only use of this function was to print `'[object Foo]'` instead of
`'[object Object]'` if an object had been created from a FunctionTemplate
with a ClassName set to `'Foo'`. With this change, we no longer set a
new toString method on the Object prototype, instead we set the
`Symbol.toStringTag` property on the prototype to the appropriate string.

PR-URL: https://github.com/nodejs/node-chakracore/pull/375
Reviewed-By: Kyle Farnung <kfarnung@microsoft.com>
Reviewed-By: Hitesh Kanwathirtha <hiteshk@microsoft.com>