[Merge Microsoft/Chakracore@7e667a05e6] [1.7>master] [1.6>1.7] [MERGE #3496 @MikeHolman] remove invalid special case in regalloc
Merge pull request #3496 from MikeHolman:regalloc

We had special case in regalloc if we are storing to a tmp reg and the next instr has the same dst, we don't need to insert a store. Unfortunately, this is invalid in case the sym is also a source of the next instr.

Benchmark perf is flat.

OS: 11480955