android: use NDK's libc++ and clang
The default NDK libstdc++ in the NDK doesn't support newer features
like std::nearbyint, which is used in V8, so we use libc++ instead.
The gcc compilers inside the NDK have a bug in the header ordering
when compiling, which causas errors relting to stddef.h , so we also
use clang instead of gcc.