[Merge Microsoft/Chakracore@0d3d21acff] [1.7>master] [1.6>1.7] [MERGE #3584 @thomasmo] OS#13419689 - DOMFastPathGetter instructions do not propagate destination profile data, adding overhead to inlined getter calls
Merge pull request #3584 from thomasmo:domfastpathgetter_dst_profile

This change addresses the issue of Get/SetElement being used a TypedArrays from the host, which should have specialized calls
from the backend. This bug happens because, when the Inliner creates the DOMFastPathGetter instruction to replace a LdFld
instruction, it does not propagate dst's profile data. Thus, though DOMFastPathGetter avoids calling into the host, it also
introduces a new cost of making generic calls because the type is unknown.
This change ensures that the original LdFld's dst profile data is also copied to the new instruction for later type-specific
optimizations.