[Merge Microsoft/Chakracore@191b97661a] [MERGE #3624 @ricobbe] OS:13383062 Don't hoist instructions out of a loop if an argument isn't primitive in the landing zone
Merge pull request #3624 from ricobbe:hoist-with-implicits-from-master

Avoid a problem, found by ExprGen, where we have an instruction in the loop body (loop condition, actually) with an argument that's tagged as a float. We're hoisting it out of the loop, despite the fact that the float tag is only valid because of an earlier instr in the loop condition, which bails if the arg isn't actually a float. As a result, in this particular test case, the hoisted instruction ends up calling the argument's valueOf method too many times, and so its side effects happen too much.