[Merge Microsoft/Chakracore@e598f2c048] [MERGE #3767 @MikeHolman] fix bug with inlining and int64 syms on x86
Merge pull request #3767 from MikeHolman:x86inlining

When splitting int64 syms on x86 we map them to a dictionary of associated int32 syms, which is on the func... but with inlining, we might have the sym in the dictionary of 1 inlinee and other inlinees won't know about it.

Fix is to always look at topFunc.

OS: 13801633