[Merge Microsoft/Chakracore@82cb2e473a] Do not assert on cross-site CustomExternalObject thrown as exception objects as they don't need marshalling
If a cross-site CEO is thrown as an exception we will not marshall it as it is an external type. But in the debugEval this cause an access denied error later when the script context is validated for the same. But we never marshall CEO's, so updating the checks accordingly.