[Merge Microsoft/Chakracore@36e9f41ca2] [1.6>1.7] [MERGE #3854 @Penguinwizzard] Unblock CI
Merge pull request #3854 from Penguinwizzard:remove_mfeds

## Remove MarkerForExternalDebugStep

This was previously used in the hybrid debugging scenario (VS would
set a breakpoint on this function, which was hard to inline). Since
this is no longer supported, we can remove this function.

## Add analysis asserts to silence false positives

Prefast started warning on some false positives. This silences them.

## Pull in Jimmy's changes to fix xplat CI

Xplat build machines now have clang3.9, we need to support that.