[Merge Microsoft/Chakracore@47bfd495e8] [MERGE #3888 @thomasmo] Introduce native unittests for FunctionExecutionStateMachine
Merge pull request #3888 from thomasmo:infinite_rejit

Introduce native unittests for FunctionExecutionStateMachine
This change introduces a set of CATCH native unittests for the state machine to emulate a handful of scenarios. This will be the first round of defense with modifying the state machine after the refactoring. In addition to tests, several stubs classes and files are added to let FunctionExecutionStateMachine compile as well as to control its behavior.
The tests in this first change include
- BasicTest
- NormalExecution
- JSUnitTestInterpreted
- JSUnitTestDynapogo
More tests will be added as more code is updated.