[Merge Microsoft/Chakracore@ca08c59b17] [MERGE #3940 @boingoing] Correctly order special symbol initialization for default arguments
Merge pull request #3940 from boingoing:SpecialNamesWithDefaultArguments

We incorrectly emit the special symbol initialization at the top of the
body scope when the function has default arguments but merged body and
param scopes. To be correct, we need to always emit the special symbol
initialization in the presence of default arguments - and always emit
those initializations before we emit the default arguments themselves.

Change here handles the case by walking the body scope var list manually
(because symbols in the body scope sym list are in the opposite order)
and emits each special symbol. We stop walking this list when we come to
the first non-special symbol as we know the parser has put all of the
special symbols at the front of the var list.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems?id=13977464&_a=edit