[Merge Microsoft/Chakracore@20ee18b364] [1.7>master] [MERGE #3945 @boingoing] Remove 'this' binding for global and global lambda functions
Merge pull request #3945 from boingoing:RemoveGlobalThisBinding

Do not create a 'this' symbol for global functions and global lambda functions. Instead emit a LdThis into a special register stored in the FuncInfo during the prolog of these functions and load that register at every name node referencing 'this'.

Add `PnFnc::IsDerivedClassConstructor` and refactor `Parser::CreateSpecialSymbolDeclarations`

`PnFnc::IsDerivedClassConstructor` is just a convenience method for `(PnFnc::IsClassConstructor() && !PnFnc::IsBaseClassConstructor())`.

`Parser::CreateSpecialSymbolDeclarations` had the code which does the check for special symbol references factored-out so it's a bit simpler.

Simplify the logic to construct the special symbol var declarations a little bit. Also fix emitting a special symbol as a call target. This fixes:
https://microsoft.visualstudio.com/OS/_workitems?id=13977724&_a=edit