[Merge Microsoft/Chakracore@9b255666b9] [MERGE #3934 @rhuanjl] Notify host Dynamic Module ready (Fixes #3902)
Merge pull request #3934 from rhuanjl:master

Previously the call to notify the Host that a module was ready was made only for:
a) Errors AND
b) Parent level modules

A module loaded with a dynamic import may not be a parent and hopefully won't always be an error - this condition change means that the notification is made for any dynamic import when completed.

Condition check used of having an attached promise was chosen to be the same as the condition used for Dynamics on the failed import path on line 194.

Fixes: https://github.com/Microsoft/ChakraCore/issues/3902