[Merge Microsoft/Chakracore@1ba3739abe] [MERGE #4072] Do not skip emitting the nested functions as the assumption is broken in VisiteNestedScopes
Merge pull request #4072 from aneeshdk:FuncBodyInitIssue
Do not skip emitting the nested functions as the assumption is broken in VisiteNestedScopes

In VisitNestedScopes while visiting a function we create new FunctionBody for all nested functions inside that. This breaks the assumption made in the PR https://github.com/Microsoft/ChakraCore/pull/3973/files. So reverting back that change. I will revisit this later for a complete fix.