[Merge Microsoft/Chakracore@e6dc0a9dea] [MERGE #4082 @akroshg] Improving perf for encodeURI.
Merge pull request #4082 from akroshg:encode

There are lots of busy-hang bugs reported for encodeURIComponent api. When I dig deeper I found that
many a times we get input larger than a meg. I created a repro test case for this. We spent most of
the time in swprintf. I made easy change to the code and got rid of that.
In the repro benchmark
Before change :
  we spent ~3500 ms.
After change :
  it becomes 450 ms.

There were opportunities to clean the code further, but I don't see a point of changing the code any more than this.