[Merge Microsoft/Chakracore@0114fa125c] [MERGE #4071 @dilijev] RegExp Bytecode: Optimize memory for Switch instructions.
Merge pull request #4071 from dilijev:rebc-switch

The number of cases supported by each switch opcode seems to have been chosen somewhat arbitrarily in the original implementation, and for the common case of few alternates, has a lot of unused space. This PR replaces the old Switch opcodes (10, 20 cases) with a new set of Switch opcodes following a pseudo-doubling scheme (2, 4, 8, 16, 24*).

\* The final increment was chosen to ensure that unused memory is never greater than the predecessor, we support at least as many switch cases as before before switching to chained layout.

This has a slightly higher cost at compile time because there are now 5 Switch opcode alternatives as opposed to the previous 2.

Memory usage is improved on average across all cases (fewer unused slots on average), especially in the presumed common case of a small number of alternatives (human edited regexps) when the switch has 8 or fewer cases. I don't yet have data about the distribution of the various numbers of cases.

# Future Work

Currently, each case takes up 2+4 [8 when padded] bytes (2 [4] for the char16, 4 bytes for the Jump target which is a uint32_t to support any size of regexp program). This could be improved with reduced layout options of 16-bit jump-targets (2+2 [4] bytes per case).

This could be further improved with a narrow layout for a single byte char for the ASCII case, and further by reducing the size of the jumps (using offset from next instruction would make this more feasible but for large numbers of cases the later targets are spaced out enough that jumps would probably not be able to fit into a single byte -- this would be improved by better packing and ASCII layout for common instructions like MatchChar).

In order to use the reduced layouts, we would have to keep track of the highest values for the various fields at compile time and use the opcode layout that best fits them.

# Example

2 cases. `/(hello|world)/`
Before (scroll to the right to note the unused slots):

```
        L0020: (0x058 bytes) SwitchAndConsume10(
            case 'h': Jump(L007c)
            case 'w': Jump(L0090)
        )
            0x000002862F6F8020[+0x000](0x004) [Inst]: 0a000000
            0x000002862F6F8024[+0x004](0x054) [SwitchMixin<10>]: 02000000 68000000 7c000000 77000000 90000000 ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff ffff0000 ffffffff
```

After:

```
        L0020: (0x018 bytes) SwitchAndConsume2(
            case 'h': Jump(L003c)
            case 'w': Jump(L0050)
        )
            0x0000022F95158090[+0x000](0x004) [Inst]: 0d000000
            0x0000022F95158094[+0x004](0x014) [SwitchMixin<2>]: 02000000 68000000 3c000000 77000000 50000000
```