[Merge Microsoft/Chakracore@3825ee2bbd] [MERGE #4069 @dilijev] RegExp Bytecode: Reduce tag size from 4 bytes to 1 byte and use #pragma pack on bytecode structs
Merge pull request #4069 from dilijev:rebc-tagsize

InstTag (opcodes) in RegExp Bytecode only need a single byte. Furthermore, there is a lot of space lost to padding in the RegExp Bytecode, so we now tell the compiler to pack the structs. This results in unaligned members of the bytecode structs, but an IB has shown this not to regress performance.

This change is the groundwork for future efforts in variable-length encoding of RegExp Bytecode.

Notably, the way we use placement new to layout bytecode in the buffer doesn't really allow the compiler to do its own alignment, so after changing the opcode size opcodes are no longer aligned. This potentially impacts perf, but @Penguinwizzard informed me that we already have this problem with the ChakraCore bytecode, and it probably doesn't hugely impact perf.

# Performance Analysis

IB shows that this does not regress perf. There were some improvements and some regressions but a follow up revealed them to be noise. If there are any regressions, they are within the noise of the measurement, and it is potentially worth it for the comparatively large improvement in memory usage.

# Memory Analysis

Using `#pragma pack(1)` on all `Mixin` and `Inst` types packs things down quite significantly. In the example below, the tag size change (4 bytes -> 1) decreased the bytecode size from 0x44 to 0x3e. `#pragma pack(1)` further reduced it to 0x29.

This change improves RegExp Bytecode memory usage by ~10% in AdBlock (reducing memory usage by 1 MB overall).

# Example

RegExp: `/(^token|foo)/`

Before (InstTag = 32 bits):

```
    instructions: {
        L0000: (0x008 bytes) BeginDefineGroup(groupId: 1)
            0x0000020E2BECC0A0[+0x000](0x004)(size:0x04)(align:0x04) [Inst]: 3b000000
            0x0000020E2BECC0A4[+0x004](0x004)(size:0x04)(align:0x04) [GroupMixin]: 01000000
        L0008: (0x008 bytes) Try(failLabel: L0028)
            0x0000020E2BECC0A8[+0x000](0x004)(size:0x04)(align:0x04) [Inst]: 57000000
            0x0000020E2BECC0AC[+0x004](0x004)(size:0x04)(align:0x04) [TryMixin]: 28000000
        L0010: (0x004 bytes) BOITest(<hardFail>: false)
            0x0000020E2BECC0B0[+0x000](0x004)(size:0x04)(align:0x04) [Inst]: 0d000000
        L0014: (0x00c bytes) MatchLiteral(literal: