[Merge Microsoft/Chakracore@e1b42728dc] [MERGE #4118 @pleath] Share types (i.e., allow PathTypeHandlers) for properties with non-standard attributes (non-writable, etc.).
Merge pull request #4118 from pleath:sharetypes-m

This change draws on Kount Veluri's work on behalf of native fields. It adds a wrinkle to the type path mechanism, replacing the PropertyId successor map key with a key that munges PropertyId and attributes. The way is paved for sharing types with getter/setters, deleted properties, and native values. When a property's attributes are changed/set, the type path is branched starting with the type that precedes the one that introduced the property.