[Merge Microsoft/Chakracore@ca7346a3db] [MERGE #4180 @sigatrev] ARM64: fix EmitUIntToFloat
Merge pull request #4180 from sigatrev:uint

In ARM32 we use different opcodes for converting ints or uints, and are not careful about the type of opnds sent to EmitUIntToFloat. In ARM64 we use the same opcode and let the opnd type decide which converison is done so we need to be sure that the right type is provided.