deps: update ChakraCore to Microsoft/ChakraCore@eca2b24e4a
[1.8>master] [MERGE #4214 @aneeshdk] Do not go through StartEmitFunction if the fuction is deferred

Merge pull request #4214 from aneeshdk:StartEmitFuncFixR1.8

There can be cases where we are reusing the function body for a child function and we got to StartEmitFunction for the child function when trying to emit the re-deferred parent function. In that case don't redo the slot allocations in StartEmitFunction for the child function as it already has valid bytecode.

Reviewed-By: Jack Horton <jahorto@microsoft.com>