deps: update ChakraCore to Microsoft/ChakraCore@0872d8f855
[1.8>master] [MERGE #4291 @jackhorton] Always disable warning 4800

Merge pull request #4291 from jackhorton:never-warn-bool

After speaking with Michael Ferris about this, we are under agreement that this is somewhat unnecessary to ever warn on. Since it only warns in 2015 and not 2017, it can also cause easy-to-miss CI failures.

See: https://msdn.microsoft.com/en-us/library/b6801kcy.aspx

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>