deps: update ChakraCore to Microsoft/ChakraCore@1b29d304cf
[1.8>master] [MERGE #4330 @MikeHolman] Change FindRegUse to HasSymUse

Merge pull request #4330 from MikeHolman:fgpeepbug

In `FlowGraph::PeepCm` we reorder code to combine CmEq+BrTrue in BrEq, and do static assignment of true/false to the result of the compare. We also do this through multiple levels of branches.

To ensure the sym we are assiging the compare result to isn't being used in intermediate instructions we call FindRegUseInRange, which checks if there are any RegOpnd uses of the sym. However, this fails to find SymOpnd uses.

````
var a = d()
a = null !== a.face && 0 < a.face.a
if (!a || b) blah = a
````
In this example, since `a` is both the dst and a src, we end up initializing `a` with `false` in the beginning of the compare and then do a LdFld on false, which throws an exception.

Other callsites seemed like they were ok, but they weren't using the RegOpnd return, so I changed the method around to return bool if there is a SymUse, which is what I think people really wanted.

Fixes #4266

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>