deps: update ChakraCore to Microsoft/ChakraCore@88ebe68372
[MERGE #4314 @MSLaguana] Add support for emitting ETW events to LTTng

Merge pull request #4314 from MSLaguana:lttng

We define a number of ETW events on windows that we can emit for diagnostics purposes. With these changes, we support generating LTTng bindings for these same events so the same diagnostics approach can be used on linux and osx.

The script that generates the LTTng bindings from the ETW manifest is based on a similar script used by the dotnet team (see https://github.com/dotnet/coreclr/blob/master/src/scripts/genXplatLttng.py), but I have attempted to simplify it and adapt it for our purposes.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>