deps: update ChakraCore to Microsoft/ChakraCore@e371362001
[1.8>master] [MERGE #4345 @sethbrenith] Look up prototype properties when running Array.prototype.copyWithin

Merge pull request #4345 from sethbrenith:copywithin

In cases where an array object is missing indices in the source data of a copyWithin operation, we need to check the prototype chain for those properties. We were calling JavascriptArray::HasItem, which examines only the current object, rather than JavascriptOperators::HasItem, which traverses the prototypes.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>