ios: rename, C API, iOS 9.0 target, parallel build
Renames libnode.framework to NodeMobile.framework . Apple framework
naming conventions declare that framework names should start with a
capital letter, and the `lib` prefix is reserved for static/dynamic
libraries.

Moves the node::Start method to a node_start C method, allowing Swift
compatibility. Swift does not support C++ interop, and importing the
libnode framework into an iOS Swift project causes a build error. To
fix this, the node::Start method was wrapped into an equivalent C
method and exposed in a C header.

Sets minimum iOS target to iOS 9.0 . There is no reason to only target
the latest version of iOS, as this library does not use new APIs
introduced in iOS 11. The target was set to iOS 9, which is a typical
minimum version that are found in apps that are backwards compatible.

Adds -j argument to make for iOS target. Adding the '-j$(getconf
_NPROCESSORS_ONLN)' flag to make in the ios_framework_prepare script
speeds up builds on multi-core machines, by building with the
available processors.

PR-URL: https://github.com/janeasystems/nodejs-mobile/pull/18
Reviewed-By: Jaime Bernardo <jaime@janeasystems.com>