deps: update ChakraCore to Microsoft/ChakraCore@1674db8744
[1.8>master] [MERGE #4464 @ricobbe] OS#14980696: Minor refactoring to fix OACR error

Merge pull request #4464 from ricobbe:build/ricobbe/bug-14980696-oacr-array-index

Refactor PathTypeHandlerBase::UpdateAttributes and UpdateSetterSlots to take the size of the new type path instead of the new type path itself as an argument.  Not only does this explicitly indicate that these methods only use the path's size and not the full path, but it also allows OACR to conclude that a couple of array element accesses are always safe.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>