deps: update ChakraCore to Microsoft/ChakraCore@a10a915d67
[1.8>master] [MERGE #4453 @thomasmo] SwitchOpt determined by AggressiveIntTypeSpec breaks repeated String cases

Merge pull request #4453 from thomasmo:release/1.8

This change fixes an issue where the function GlobOpt::IsSwitchOptEnabled was updated to add conditionsfor type specialization. This resulted in bad IR being produced, which impacted how type-specialized BailOutswere constructed and which registers were restored (or, not restored).
The fix is to apply align optimizationchecks in SwitchIRBuilder based on the types of the cases.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>