deps: update ChakraCore to Microsoft/ChakraCore@6311ef5a45
[MERGE #4529 @boingoing] Add JSRT support for native callbacks with `new.target`

Merge pull request #4529 from boingoing:NativeCallbackSupportNewTarget

Existing `JsNativeFunction` callbacks have no concept of `new.target` and, as a consequence, native functions cannot be used as the base class for a derived class constructor in Javascript. An API like `JsGetNewTarget` doesen't make a lot of sense so we propose here a new type of callback which will be called with `new.target` directly.

`JsCreateEnhancedFunction` acts similarly to the way `JsCreateFunction` did but takes a `JsEnhancedNativeFunction` callback instead. This callback is called passing a struct (`JsNativeFunctionInfo` containing `new.target`, `this`, and `isConstructCall`) as well as ordinary `callee`, `arguments`, and `argumentsCount`.

To simplify `JavascriptExternalFunction` and handling of native callback functions in ChakraCore, I refactored `JsCreateFunction` to use a `JsEnhancedNativeFunction` wrapper function which forwards to the `JsNativeFunction` callback.

Add native unit tests where an enhanced function (`JavascriptExternalFunction`) is used as the base class for a Javascript derived class.

Fixes: #3762

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>