deps: update ChakraCore to Microsoft/ChakraCore@a0d8bdd37e
[1.8>1.9] [MERGE #4565 @aneeshdk] Fixing the cross context behaviour in DebugEval

Merge pull request #4565 from aneeshdk:DebugEvalCrossContextIssue

DebugEval method is invoked in debug eval script context. Then the string that user enters is executed in the target script context. The newly created function object belongs to target script context but the current top of the stack is pointing to the debug eval script context. We have to marshall the function object before invoking it to fix the stack.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>