deps: update ChakraCore to Microsoft/ChakraCore@d8c169fbb0
[1.9>master] [1.8>1.9] [MERGE #4628 @boingoing] Defer parse treats yield as keyword

Merge pull request #4628 from boingoing:DeferParseYieldToken

When we're defer parsing a function, we construct an enclosing function context to use as a parent function. That dummy parent function does not initialize the function flags, though, so they're random. When we're parsing a deferred lambda function, we look at those flags in the dummy parent function to see if it was a generator function to know if we should treat yield as an identifier or not. Because the flags are random, we sometimes throw a syntax error here thinking yield should be treated as a keyword. If we throw a syntax error upon defer-parse, we will hit an assert.

Fix by storing a bit of state in the ScopeInfo which indicates if the enclosing function is a generator or is an async function.

Fixes:
https://microsoft.visualstudio.com/web/wi.aspx?id=14502906

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>