deps: update ChakraCore to Microsoft/ChakraCore@6802d8248b
[1.9>master] [1.8>1.9] [MERGE #4629 @boingoing] OS#15334058: Fix `class extends null` and `this` access for super property reference

Merge pull request #4629 from boingoing:ClassExtendsNull

We were supporting some old version of the spec language around classes which extend from null. Instead of leaving `this` undeclared, we were constructing an object and assigning it the `this` binding. This meant we were not throwing ReferenceErrors on access to `this` when we should have been.

Also fixed the case where a super property reference failed to access `this` resulting in trying to load a property from null.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/15334058

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>