deps: update ChakraCore to Microsoft/ChakraCore@a7f6ef7b3e
[1.9>master] [1.8>1.9] [MERGE #4618 @boingoing] OS#14568840: Remove 'this' binding for indirect eval

Merge pull request #4618 from boingoing:RemoveThisBindingIndirectEval

Having a 'this' binding in the indirect eval leads to problems if there is a lambda capturing 'this' in the indirect eval. The lambda would try to load 'this' from a scope slot in the global scope of the indirect eval which asserts.

Seems we can simplify the above by just removing the 'this' binding from the indirect eval. Then we'll simply load 'this' like an ordinary lambda at global scope would.

Fixes:
https://microsoft.visualstudio.com/web/wi.aspx?id=14568840

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>