deps: update ChakraCore to Microsoft/ChakraCore@97d5c20304
[1.9>master] [1.8>1.9] [MERGE #4612 @boingoing] OS#13976524: Force scanner to create pid for string constant trailing defer-parsed compact-body lambda

Merge pull request #4612 from boingoing:13976524

Lambda with compact-body which is defer-parsed and which is trailed immediately by a string constant can cause the string constant to miss having a pid created. When we try to load the string constant table in `ByteCodeGenerator`, we'll hit an AV with this pid being nullptr.

Fix by forcing the scanner to create a pid for this string constant - only for the single case of a string constant token immediately following the single-expression lambda body which is defer-parsed.

Test case looks like this:
```
jtmchw => z
'123'
```

Fixes:
https://microsoft.visualstudio.com/web/wi.aspx?id=13976524

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>