chakrashim: Fixing v8::ArrayBuffer implementation detail
v8::ArrayBuffer was returning raw internal pointers to ArrayBuffers.
This is fine in general, but in v8 in particular these pointers are
guaranteed to be non-null. For chakracore in the case of empty ArrayBuffers
this implicit assumption was broken, which showed up in one test scenario

We now explicitly check for this case and return a valid (if useless) pointer
instead.