deps: update ChakraCore to Microsoft/ChakraCore@a48d5fcab3
[1.9>master] [1.8>1.9] [MERGE #4640 @obastemur] stack/error: disable it partially for xplat CI

Merge pull request #4640 from obastemur:dis_stack_test_xp_ci_partially

reduces test run time to 1/3

I've got bunch of CI failures due to this test is timing out (xplat). disabling a part of it helps to save it from whole run is going timeout.

I did look into the details and saw `interlocked_increment` was playing a big role on the timeout. CH loses its' share from OS scheduler.. If CI machine is busy, it takes UNKNOWN amount of time for VM to get back the share needed.. well it loses again very soon because of the same reason.

So, I won't add (todo, enable this) or open an issue to track this.  Also, see the comments on the test file.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>