test: fixing test failures
A couple of tests had incorrect baselines,
test-tty-get-color-depth incorrectly tried to get the color of
stdin, and test-fs-write used a new v8-specific flag.

To solve that last case, I added a list of banned flags to
the python script that runs tests, which will cause it to ignore
those flags if they are requested by a script.