deps: update ChakraCore to Microsoft/ChakraCore@e4e4805dc7
[1.9>master] [1.8>1.9] [MERGE #4662 @tcare] Fix #2949: Class const name binding incorrectly marked as module export storage

Merge pull request #4662 from tcare:classexport

We were mistakenly marking both the class let and const decls as module export storage. This resulted in empty instantiated scopes and incorrect class behavior in some cases.

The fix is not to mark the const binding as module export storage, as it is visible to the inside of the class only.

Added some tests to verify the bindings of exported classes.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>