deps: update ChakraCore to Microsoft/ChakraCore@60c7905ad2
[1.9>master] [1.8>1.9] [MERGE #4666 @leirocks] fix missing function table clearing for loop body cleanup

Merge pull request #4666 from leirocks:xdata1

 clear the function table for loop body entrypoint case.
when debugger attaching the code address is not freed through normal path hence clearing the xdata in main thread. this is done for function entrypoint in last change but missed loop body case. also make sure of the integrity of clearing function table in the slist.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>