deps: update ChakraCore to Microsoft/ChakraCore@c6c0ff7ef9
[1.8>1.9] [MERGE #4673 @boingoing] OS#14115684: Function-in-block semantics don't work for defer parse

Merge pull request #4673 from boingoing:DeferFunctionInBlock

In below code repro, we do not create var decl binding for `foo` inside the `nest` function in defer-parse mode. Due to this, we leave the reference to `foo` in `nest` on the pid ref stack and bind it to the formal `foo` parameter from the `test` function. This leads us to think formal `foo` is captured and try to put it into a scope slot. However, formal `foo` is not actually captured so no scope slot is allocated. When we actually generate bytecode for `test`, we hit a fail fast.

```js
function test(foo) {
    function nest() {
        {
            function foo() {
                console.log('pass');
            }
        }
        foo();
    }
    nest();
}
test(()=>console.log('fail'));
```

Fix is to create the var binding for `foo` at the function declaration for `foo` even in defer-parse mode.

Fixes:
https://microsoft.visualstudio.com/web/wi.aspx?id=14115684

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>