deps: update ChakraCore to Microsoft/ChakraCore@54e986f8a4
[1.8>1.9] [MERGE #4682 @MikeHolman] Fix bad interaction with Spectre mitigation and VirtualArray OOB resume

Merge pull request #4682 from MikeHolman:virtualspectre

In case of VirtualArrays, we may have eliminated bound check and rely on our AV handling (as long as index is guaranteed to be within 4GB).

However, with spectre mitigations we force OOB reads to nullptr. Our exception filter only handles AVs trying to read from the reserved region, so we end up crashing with nullptr deref instead of resuming.

This change makes it so that we will only poison in case the index exceeds our 4GB reservation.

OS: 15897366

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>