deps: update ChakraCore to Microsoft/ChakraCore@81b7667c4e
[MERGE #4709 @sethbrenith] Fix a recent regression where we could hit an assert when attempting to recursively inline a function that came from dynamic script.

Merge pull request #4709 from sethbrenith:user/sethb/dynamic-call-inline

I recently removed a check for matching secondary host source context when deciding whether to inline a getter or call/apply target. In certain circumstances, that change can cause an assertion to fail, so this change adds back the check.

The assertion in question is in `DynamicProfileInfo::GetCallSiteInfo`, and it asserts that the function must either be built-in, a public library function, or `HasCallSiteInfo(functionBody)`. `HasCallSiteInfo` checks whether the function body's source context matches the `noContextSourceContextInfo`, which is the case for functions generated from strings. However, if I ignore the assertion and continue, `callSiteInfo` seems to exist and contain correct data, which leads me to think that `HasCallSiteInfo` might be overly restrictive. However, the safe change is just to block this particular inlining case.

Fixes OS:15987143

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>