deps: update ChakraCore to Microsoft/ChakraCore@bdabd87d40
[MERGE #4726 @pleath] Refactor PathTypeHandler successor info.

Merge pull request #4726 from pleath:succinfo

Simplify the hierarchy of PathTypeHandler subclasses by factoring the set of successors into a separate structure. Creates PathTypeSuccessorInfo (with subclasses for single and multiple successors). Eliminates SimplePathTypeHandler and friends, as well as the PathTypeHandler class that's currently the common parent of the instantiable multi-successor classes.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>