deps: update ChakraCore to Microsoft/ChakraCore@81aaed23dc
[MERGE #4731 @leirocks] fixes for dynamic profile caching

Merge pull request #4731 from leirocks:dpcmaster

Previously there's leak because dynamic profile info is pinned and causes two roots(another is global function body) on a ring.
The fix is to use a heap allocated structure to save the information to dump to dynamic profile cache.
The fix missed a scenario the while loading from cache, it didn't save to the heap allocated dynamicProfileInfoMap,
and later some function is inlined, cause the info is missed while saving to the cache file, causes the cache file mismatch with the script code

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>