deps: update ChakraCore to Microsoft/ChakraCore@939118df9f
[MERGE #4769 @boingoing] OS#15679026 Fix incorrect type check in OP_SetProperty

Merge pull request #4769 from boingoing:FixSuperReferenceNullInstance

`OP_SetProperty` takes an optional `thisInstance` parameter which is used only in the super property reference case because we need to maintain 'this' binding for class member accessors. Problem is when `thisInstance` is provided, we check the type of it instead of the type of the receiver object to determine if the property assignment is invalid due to the receiver object being null, undefined, etc.

This has the unfortunate side-effect of forcing us to try and do the property set when the receiver is null which results in an assert down in the CacheOperators because they assume that the receiver is an object with a `DynamicType`. We can only hit this code path when we have a super property reference in a class member of a class which extends from null.

```
class c extends null {
    constructor() {
        return {}
    }
    meth() {
        super['prop'] = 'something';
        super.prop = 'something'
    }
}
c.prototype.meth.call({})
```

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/15679026

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>