deps: update ChakraCore to Microsoft/ChakraCore@99918dc823
[1.9>master] [1.8>1.9] [MERGE #4785 @MikeHolman] Different CRTs declare ceil/floor with different calling conventions, so use our own

Merge pull request #4785 from MikeHolman:crtcdecl

Previously we were assuming cdecl. And while this was true for floor, it was not true for floorf in the CRT version that ChakraFull uses. However, on the version of CRT I'm using with ChakraCore, floorf is cdecl.

Because the calling convention is not reliable, I added methods in chakra for the JIT to call to ensure the calling convention is what we expect (at least for MSVC builds).

Also, since it's our function now, we can use callee cleanup.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>