deps: update ChakraCore to Microsoft/ChakraCore@f3c5f9f215
[MERGE #4748 @pleath] Share SimpleTypeHandler types on function objects.

Merge pull request #4748 from pleath:sharesimpletypes

Currently when we transition from DeferredTypeHandler to SimpleTypeHandler on first access to a function object's properties, we don't share the new type we create. Track an undeferred type along with the deferredPrototypeType on FunctionProxy and re-use it whenever possible for multiple instances of the same function. (This is a first step in broader sharing of function object types.)

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>