deps: update ChakraCore to Microsoft/ChakraCore@d786d4c6e4
[MERGE #4819 @sigatrev] Optimize In operator (OpCode::IsIn) for int indices in arrays

Merge pull request #4819 from sigatrev:isIn

This commit adds a globopt optimization path to turn IsIn into a LdTrue for arrays when the the index is an int, the array has no missing values, and the index can be guaranteed to be within bounds. It leverages the existing bounds check elimination code for LdElemI/StElemI.

The motivation for this optimization is to improve efficiency for JavaScript implementations of built-in functions.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>