deps: update ChakraCore to Microsoft/ChakraCore@722435a628
[MERGE #4834 @sethbrenith] Speed up `in` operations with PropertyString or Symbol keys, part 1

Merge pull request #4834 from sethbrenith:user/sethb/property-string-in

This change allows the runtime call for `str in obj` to use the same property record cache that we use for loading `obj[str]`. This makes the `in` operator faster in some common cases, and also populates the cache for an upcoming change which will lower fast-path code for `in` on PropertyStrings and Symbols.

With this change, the Elm test in Speedometer improves about 3%, and time for the following microbenchmark drops from about 440 ms to about 240 ms:

```javascript
function compare(a, b) {
    let diffs = 0;
    for (const aKey in a) {
        if (!(aKey in b)) {
            ++diffs;
            continue;
        }
        if (a[aKey] !== b[aKey]) {
            ++diffs;
            continue;
        }
    }
    for (const bKey in b) {
        if (!(bKey in a)) {
            ++diffs;
            continue;
        }
    }
    return diffs;
}

const before = {
    a: 1,
    b:

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>