deps: update ChakraCore to Microsoft/ChakraCore@38425ca4bf
[MERGE #4863 @agarwal-sandeep] Remove isEval field from ByteCode Symbol class

Merge pull request #4863 from agarwal-sandeep:symboliseval

Everytime a Symbol is created it construct another symbol for eval and
then compare to check if current symbol is eval. This shows up in profile
traces during defered parsing when ExtractScopeInfo is called. This field
is never used so remove it and save some work.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>