deps: update ChakraCore to Microsoft/ChakraCore@8937b1a8fd
[MERGE #4866 @VSadov] Be more conservative about Mul_A operand values in prepass to avoid rejit.

Merge pull request #4866 from VSadov:Investigate4744

Fixes:#4744

JIT tries to typespec `result` in the following code to `int32`, even with AggressiveMulIntTypeSpec off

That later results in a JIT time bailout when `result` fails to loslessly convert and triggers rejit with AggressiveIntTypeSpec  off.

Interestingly, in this particular scenario, it happens only if `result` is initialized to literal `1`.
If it is initialized to `2` or even to `1` , but in a more roundabout way, the code runs more than 2 times faster, since it becomes a float Mul right away.

```JS
function factorialA(n)
{
    let result = 1;
    for(let i = 0; i < n; ++i)
    {
        //   1 * whatever     in prepass
        //  seems like will not overflow, right?    not really...
        result *= (n - i)
    }
    return result;
}
```

The fix is to be less aggressive with no-overflow predictions for Mul_A when AggressiveMulIntTypeSpec is off

The idea is - If we don't definitely know the operand's min/max values, or if the operand is live on the back edge, we can assume it will have full range and will likely overflow and should just use float math, unless the other operand is definitely very small. -  basically `1` or `0`. Even `result *= 2` is likely to overflow real quick since growth, if happens, is exponential.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>