deps: update ChakraCore to Microsoft/ChakraCore@1a6bcac0c7
[MERGE #4943 @rajatd] Making object pointer copy-prop (slightly) more agressive in loop prepass

Merge pull request #4943 from rajatd:safeToTransfer

Object pointer copy prop in pre-pass will only kick in if both the object sym and its copy prop sym are single-def. I'm changing it such that the we'll do the optimization in prepass if -

both object sym and copy sym should not be live on back-edge, if they are then they should not be assigned to inside the current loop or any of its parent loops.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>