deps: fix libuv registry API error handling
The `Reg*` APIs on Windows don't use `GetLastError` to report failures.
The errors are returned directly from the call.

For systems which don't have one of the values `GetLastError` can end
up returning `0` to the caller, indicating success. The caller then
assumes that the data is valid and can attempt to execute on garbage
data. This change fixes the flow to correctly return the error to the
caller.

Upstream-PR: https://github.com/libuv/libuv/pull/1811
PR-URL: https://github.com/nodejs/node-chakracore/pull/517
Reviewed-By: Seth Brenith <sethb@microsoft.com>