deps: update ChakraCore to Microsoft/ChakraCore@4eef2f4495
[MERGE #5010 @jackhorton] Support Chakra.ICU on ARM

Merge pull request #5010 from jackhorton:icu/arm

This is the next major step in converting to ICU on all platforms. This change converts Chakra.ICU.Toolutil and Chakra.ICU.GenCCode to be built to run on the host architecture (usually x64) in order to generate icudtl_dat.obj, which is platform-agnostic and can be linked into an ARM-compatible DLL using the regular ARM linker. As a downside, it means that a clean build of Chakra.ICU takes about twice as long because ICU's common and i18n code are built twice, once for the target that gets linked to the resulting ChakraCore.dll, and once for the host that gets linked to GenCCode.exe. This is going to be a major problem for CI, however we can (hopefully) solve that problem later through caching

/cc @Microsoft/target-dev (I cant add target-dev as a reviewer for unknown reasons). I think this will also enable ARM64, however I haven't checked that yet.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>