deps: update ChakraCore to Microsoft/ChakraCore@08cf998b1d
[1.8>1.9] [MERGE #5074 @kfarnung] Breakpoint APIs don't verify current context

Merge pull request #5074 from kfarnung:diagbp

The JsDiag APIs for breakpoints use the `GlobalAPIWrapper_NoRecord`
wrapper which does not validate that there's a current context. During
the function they then get the current context and attempt to use it
without checking whether it's `nullptr`.

The fix is to use `ContextAPIWrapper_NoRecord` instead which will
ensure that there's a current context.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>