deps: ensure that uv handles are cleaned up before freeing them
Our IsolateShim has a few libuv handles in it which form part of
a list of handles. When the shim is freed, we close the handles,
but without running the UV loop they don't get properly cleaned up
and so we could run into use-after-free situations.

This is unlikely to have impacted normal node scenarios since we
do not support multiple isolates at the same time, but it impacts
cctest which creates several isolates in sequence.