deps: update ChakraCore to Microsoft/ChakraCore@3fd052ee7f
[MERGE #5036 @rhuanjl] Restore SuperRestrictionState after methods fix #5030

Merge pull request #5036 from rhuanjl:fixSuper

Wanted to learn more about how parsing works so thought I'd have a look at #5030

And unless I've got this wrong the fix is very simple.

When parsing a class member or a function the SuperRestrictionState is restored using AutoParsingSuperRestrictionStateRestorer after parsing the member or function is complete.

This was not being done for getters/setters or methods within object literals. This had not previously been noticed as it's only relevant for getters/setters/methods that are declared in an object literal inside a class constructor before super is called - a somewhat niche case I assume.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>