deps: update ChakraCore to Microsoft/ChakraCore@a6a3a40afb
[MERGE #5118 @thomasmo] Add support to clean up BGParse requests

Merge pull request #5118 from thomasmo:bgparse_discard

This change updates BGParseManager to allow for interruptions at various points of the background parse process. A new function, DiscardParseResults, will take the appropriate actions if the workitem is queued, processing, or processed. If the workitem is queued or processed, it will simply be removed from the appropriate lists and the items are freed. If the workitem is processing, it cannot be freed immediately, so it is marked as 'discarded' so that it can be freed at a safer time after it has been processed. While in the discarded state, the processing workitem will own freeing the buffer for the script source since the caller is no longer interested in it and the memory must be valid for background parse to complete.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>