deps: update ChakraCore to Microsoft/ChakraCore@403636dbc0
[MERGE #5143 @dlibby-] Ensure that finalizeCount on SmallRecyclerVisitedHostHeapBlockT is updated in all builds

Merge pull request #5143 from dlibby-:build/dlibby/fix-recyclervisited-finalizeCount

The incrementing of finalizeCount was being done inside an ifdef that is not enable on release
builds. This can result in finalizable objects not getting disposed properly since there are checks in
Sweep that try to release pages if there are no finalizable objects in the block.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>