deps: update ChakraCore to Microsoft/ChakraCore@bf38e2a607
[MERGE #5135 @irinayat-MS] Mark strings used by 'in' operator as property names

Merge pull request #5135 from irinayat-MS:PropString_In_Operator

Micro optimization to create PropertyStrings for strings that are known to be property names (because they are being used as such in 'in' operator). Targeted benchmarks show x2 improvement (runtime), telemetry from crawler registered no difference but also showed that the optimization's trade off of using more memory isn't a concern (same amount of strings is registered as properties with or without the optimization). Thus, the fix has a potential of improving a subset of specific usage patterns without any known broard drawbacks, so it was decided to take it.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>