deps: update ChakraCore to Microsoft/ChakraCore@249b4af2ba
[MERGE #5213 @sigatrev] OS#17384939: avoid race condition when writing callback info IDL

Merge pull request #5213 from sigatrev:callbackrace

original setup would add elements to the front of the list when profiling instructions, and would iterator over the whole list while writing the IDL. If an element was added between creating the IDL array and beginning the iteration, too many elements would be written and we'd AV.

This changes the setup to add new elements to the end of the list, and limit the number of iterations based on the size of the IDL array created. This should allow elements to be safely added to the list after creating the array. Elements are never removed from the list.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>