deps: update ChakraCore to Microsoft/ChakraCore@95e1c2b13f
[MERGE #5252 @rajatd] Fixing obj-ptr copy-prop in prepass. OS 17555305

Merge pull request #5252 from rajatd:ac-master

For copy-propping in prepass, it is not sufficient that the two syms are safe to tranfer values.  For instance, the copy sym may not be live on back-edge but, may become live on back edge after the copy-prop.

In addition to the safe-to-transfer checks, the copy-sym shouldn't be written to if the original sym is live on back edge. If, however, the original sym is not live on back edge, then we can do the copy prop, provided both symbols satisfied the safe-to-transfer checks.

Example:

    s1 = s2
    $Loop:
        s3 = s1  <-- can't copy-prop s2 here even though both s1 and s2 are safe to transfer
        s2 = s4
        Br $Loop

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>