deps: update ChakraCore to Microsoft/ChakraCore@c9a6482126
[MERGE #5257 @boingoing] When deserializing the parser state cache, verify the script code matches the script used to create the cache

Merge pull request #5257 from boingoing:crc_parserstatecache_master

Calculate a checksum of the script code and store it in the stream along with the parser state cache. When we try and use this parser state cache later, make sure the script we are given back matches the script we used to create the cache. If it doesn't for some reason, discard the parser state cache as it may not match the script code exactly.

Fixes:
https://microsoft.visualstudio.com/OS/_workitems/edit/17542547
https://microsoft.visualstudio.com/OS/_workitems/edit/17713429

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>