deps: update ChakraCore to Microsoft/ChakraCore@d63bb992c4
[1.10>master] [MERGE #5395 @paolosevMSFT] OS#17614914 - SCCLiveness::ProcessStackSymUse lifetime nullptr deref

Merge pull request #5395 from paolosevMSFT:17614914

In bug 17614914 JIT is causing a bad code gen, but the root problem is that with a script like:

```
Object.prototype.length = undefined;
var ary = new Array();
var func0 = function() {
    for (var _strvar2 in ary) {
        console.log(_strvar2);
    }
};
func0();
```

a built-in property like Array.length should not be enumerated even if it is defined in Object.prototype.
This PR fixes this problem by ignoring the builtin properties in ForInObjectEnumerator::MoveAndGetNext().

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>