deps: update ChakraCore to Microsoft/ChakraCore@4240b1d61b
[1.10>master] [MERGE #5433 @VSadov] Fixing possible overflow in BoundFunction::NewInstance.

Merge pull request #5433 from VSadov:oacr1

Not sure if the possibility is practically reachable, but the static checker complains about a  possible overflow.

We allocate an extra element in `newValues` based only on `CallFlags_ExtraArg` flag. Therefore we should be consistent when writing to that element.

`args.HasExtraArg` looks at more flags and, in theory, could result in writing to an element that does not exist. That seems to be making OACR unhappy.

Fixes: OS:#17999665
Fixes: OS:#18010300

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>