deps: update ChakraCore to Microsoft/ChakraCore@05d5c4b72d
[MERGE #5413 @Penguinwizzard] Add dump:assembly for close-to-assemblable output

Merge pull request #5413 from Penguinwizzard:testtrace_assemblable

We got info about a use case that may benefit significantly from closer-to-machine-assembly output, so doing -dump:assembly with these bits should provide something closer to that. Note that there's still some symopnds in the output, and it doesn't necessarily support every case, but it should be enough for simple analysis. The most complicated part of the change is adding support for what's nearly a two-column mode to Output.

Edit: changed flag to new version.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>