deps: update ChakraCore to Microsoft/ChakraCore@de80038b6c
[1.10>master] [MERGE #5448 @tcare] Add JavascriptNumber fastpath in JavascriptOperators::GetIndexTypeFromPrimitive

Merge pull request #5448 from tcare:index

Mitigates OS: 17348829.

We recently reactivated the JS Built In implementation of indexOf. It seems there is some bad codegen in the JIT that is passing an unsigned integer as a double through this path, causing a slow ToString of the index integer and causing a significant perf hit.

Regardless of the JIT issue, there's no reason for us to do a ToString on a JavascriptNumber if we don't have to. Added a check to see if we can use it as an index, otherwise go through the ToString as usual.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>