deps: update ChakraCore to Microsoft/ChakraCore@b371ceb2cc
[1.10>master] [MERGE #5451 @paolosevMSFT] Ensure that GeneratorReturnExceptions are not handled by InterpreterStackFrame::DebugProcess

Merge pull request #5451 from paolosevMSFT:user/paolosev/bug17418795

When running a generator function with the -debuglaunch flag the engine crashes with simple code like:
```
  function* test() {
    try {
      yield 42;
    } catch (e) {
    }
  }
  var t = test();
  t.next();
  var v = t.return();
```

The reason is that a special exception is thrown by JavascriptOperators::OP_ResumeYield() to carry the return value from a generator. When the generator has a try/catch, like in this case, this exception is caught and rethrown by InterpreterStackFrame::OP_TryCatch(), and after that it is caught and handled by JavascriptGenerator::CallGenerator().

But when we run with -debuglaunch the exception gets caught by InterpreterStackFrame::DebugProcess(), which does not rethrow it. Therefore the generator code does not have a chance to complete correctly and the process crashes because an invalid var is returned in place of a generator yielded object.

As fix this PR adds a check for IsGeneratorReturnException() in InterpreterStackFrame::DebugProcess() to avoid skipping this exception.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>