deps: update ChakraCore to Microsoft/ChakraCore@d07a94270a
[1.10>master] [MERGE #5472 @MSLaguana] Removing write barrier on non-recycler data

Merge pull request #5472 from MSLaguana:fixSerializeAccess

When detaching a `JsrtExternalArrayBuffer` we heap allocate a `JsrtExternalArrayBufferDetachedState`, but on this non-recycler object we had a `Field` which would attempt to set the dirty bit in the card table.

Because the object wasn't recycler allocated, the index is out of bounds of the card table, leading to an access violation.

Fixes #5461

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>