deps: update ChakraCore to Microsoft/ChakraCore@d86452259d
[MERGE #4837 @Penguinwizzard] Improve compatibility with Clang for Windows

Merge pull request #4837 from Penguinwizzard:windows_clang

This changeset removes some of the impediments to building ChakraCore with clang on windows. This is still not a supported configuration, just a less buggy one. There's still known failures with the staticinterpreterthunk in this configuration, and building the jit idl takes manual configuration of the relevant project file due to a clang bug.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>