deps: update ChakraCore to Microsoft/ChakraCore@8dae18aacf
[1.10>master] [MERGE #5476 @Penguinwizzard] Fix a set of permissive- issues

Merge pull request #5476 from Penguinwizzard:permissiveminus

This is a set of fixes for some of the issues that permissive- hits in our codebase. There's a few more (~15 or so right now) that I didn't get around to yet, and won't address in this PR (they tend to look like they need more extensive changes to address).
Fixes #4596
Fixes #5189

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>