deps: update ChakraCore to Microsoft/ChakraCore@1158355b25
[1.10>master] [MERGE #5497 @irinayat-MS] Move re-throw of exception out of catch

Merge pull request #5497 from irinayat-MS:StackOverflow

Fixes OS.18246154

Re-throw from catch doesn't unwind the stack (only resets the active frame pointer) so, if there are nested try/catch blocks that refrow, the stack keeps growing, which is unhelpful if the initial exception is trying to prevent Stack Overflow. To avoid this, move re-throw outside of the catch (so the stack fully unwinds on each handling site).

Took the opportunity to add more test coverage for typeof and fix exception handling deviations from the spec (and V8).

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>