deps: update ChakraCore to Microsoft/ChakraCore@a2aa450af0
[1.10>master] [MERGE #5535 @rajatd] Don't add compensation code for array values in prepass. OS#17527968

Merge pull request #5535 from rajatd:avt-array

With aggressive transfer of values in loop prepass, an assignment of an array will result in an Array kind value for the dst (without value transfer in prepass, it would have been a Generic kind value).

Globopt, when trying to merge two Array kind values from two edges, will try to insert compensation code if, for example, the length sym on the two values is different. We are currently hitting an assert that we should only be inserting compensation code in the main pass of the loop. Before AVT, we wouldn't even have tried to insert compensation because the value types on the two edges would be different.

Given the timeframe and release being shutdown, I'm changing the code so it only inserts compensation code in the main pass (instead of inserting compensation in prepass). This matches the behaviour to what we had before AVT.

Thanks @meg-gupta for the reduction on this one.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>