deps: update ChakraCore to Microsoft/ChakraCore@80bceff351
[1.10>master] [MERGE #5547 @meg-gupta] OS#15997192: Disable strict valuetype assert for hoistable syms

Merge pull request #5547 from meg-gupta:disableassert

The assert says that for a hoistable value, if its valuetype is primitive in the loop landing pad, it should be atleast LikelyPrimitive inside the loop.
This assert does not hold true in a few cases when we force specialize syms in the loop landing pad.
Ex :

  landingPad:
    a = v1;
    b = v1;
  loop :
    b = v2; // forces float conversion in landing pad
    c = a;
  jump loop

During force specializing in loop landing pad of b, both a and b have the same valueinfo, so they transition to a DefiniteFloat.
This takes b to definite float in the loop landing pad but not inside the loop.
We do not have a way to force Specialize a in the loop, we cannot assume force specializing b again in the loop will update the valueInfo of a,
because they may have different value numbers in the loop. Since we don't have a way to lookup values by valueNumber, there is no easy way to fix this gap currently.

Added test to keep track of this gap.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>