deps: update ChakraCore to Microsoft/ChakraCore@4f6b81be52
[1.10>master] [MERGE #5528 @Cellule] Post Op Bailout before first instr

Merge pull request #5528 from Cellule:users/micfer/handlerscope

Handle cases where we try to bailout before the first bytecode instr.
OS#17686612
Right now, it is possible to have a post-op bailout on LdScopeHandler which is added in IRBuilder.
I am not sure how to write a test that triggers this path, it seems specific to browser/node scenario

I have checked with @rajatd that a bailout there is fine since we will re-execute the code in the bailout path (more specifically in the first iteration of the interpreter).

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>