deps: update ChakraCore to Microsoft/ChakraCore@07271b96d2
[MERGE #5655 @wyrichte] Fixes #5532 null pointer dereference in EmitBooleanExpression

Merge pull request #5655 from wyrichte:bug_5532

Fixes #5532
Given a nested for loop that qualifies for loop inversion, the lack of a conditional in the outer loop causes a read access violation when the conditional’s ParseNode’s nop is read. The conditional is analyzed to perform a zero trip test which can skip the execution of the loop entirely. In the case where the outer loop lacks a conditional, the zero trip test will not pass. Therefore when the outer loop lacks a conditional skipping the inclusion of a zero trip test is valid and avoids the read access violation.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>