deps: update ChakraCore to Microsoft/ChakraCore@6426b607d2
[MERGE #5658 @rhuanjl] Add check for non-ascii fix:#4268

Merge pull request #5658 from rhuanjl:ascii

This PR adds an additional test to CI to block non-ascii characters, the tests, tools and Build folders are excluded from the check.

This is based on the eol check test but with a different grep command.

Additionally I found some non-ascii at the start of JavascriptString.cpp so have removed it.

@dilijev please can you take a look

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>