deps: update ChakraCore to Microsoft/ChakraCore@6a9b83bbb4
[MERGE #5716 @sigatrev] fix two bailout issues with callback.call inlining

Merge pull request #5716 from sigatrev:callback

First, there was no check to ensure .call/.apply were the real .call/.apply. There is now a CheckFixedFld for loading .call/.apply.

Second, when bailing out on the callback function not equal to what was inlined, the register for the .call/.apply method was not being restored.

Reviewed-By: chakrabot <chakrabot@users.noreply.github.com>